+++ RELEASE_NOTES.md
@@ -59,6 +59,7 @@
 - After running an autofocus run a new json file will be written to %localappdata%\nina\autofocus containing all datapoints to retrace what was measured at a later point in time
 - Reintroduce a "No Guider" option under Equipment -> Guiding so that unguided setups do not suffer through unwanted PHD2 executions or unnecessary errors when the "Connect All Devices" button is pressed.
 - Platesolve tab inside Imaging Tab will now retain its settings (Sync, Reslew, Repeat)
+- Added the option to retreieve the coordinates of the center of the sky chart instead of the selected object when using TheSkyX as the planetarium
 - Added ability to record flat exposure times with the flat wizard, to be used in a sequence later
 
 ## Bug fixes